+++ src/providers/query.provider.tsx
@@ -1,10 +1,19 @@
+"use client";
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 import { ReactQueryDevtools } from "@tanstack/react-query-devtools";
-import { PropsWithChildren } from "react";
-
-const queryClient = new QueryClient();
+import React, { PropsWithChildren } from "react";
 
 const QueryProvider = ({ children }: PropsWithChildren) => {
+  const [queryClient] = React.useState(
+    () =>
+      new QueryClient({
+        defaultOptions: {
+          queries: {
+            staleTime: 60 * 1000,
+          },
+        },
+      })
+  );
   return (
     <QueryClientProvider client={queryClient}>
       {children}
